requirements.txt
fastapi
uvicorn
pymongo
python-multipart
firebase-admin
python-docx
pypdf
pdfminer.six
scikit-learn
numpy
pymupdf
sentence-transformers
torch